MNT/FIX: Update wrappers with proper argument order
The Python side expects (date, lon, lat, z), which is what the
Fortran subroutines is defined as, but the wrapper interface defintion
had lon/z swapped (corresponding to the actual Fortran implementation).

This doesn't affect any results when using run since the interface uses
the argument order as inputs and they were all the same type. It does
affect calling the pygtd7d/pymsiscalc routines directly with keyword
arguments though as those weren't in the right order.

Added a test to make sure the keyword arguments are working as expected now.