FIX: Remove obvious solar radio burst F10.7 data
There are some anomalous F10.7 values over 400 which were happening
during a solar radio burst and don't allow the model to run when values
are over 750 (MSIS returns nans for everything after that point).
For now, lets flag values over 400 and fill it with F10.7a values
on that same date. We can change this later if there is a more definitive
way to justify these values, but this at least allows users to run over
the entire dataset without MSIS crashing.