FIX: Handle updating options without changing other variables
There was an issue in MSIS2 where some steps are cached in the
Fortran code if the input values (lat, lon, ...) are unchanged
between subsequent runs. This causes issues if one wants to compare
what the differences are between the options switches. (MSIS00 didn't
have this same cache)

Additionally, add a check to the Python side to only call the init()
function when options have changed, not every time we enter the run()
call.