ENH: Add multithreaded support and locking for freethreaded builds
This adds some locks around the calls to make sure we are only entering
the Fortran routines from a single thread at a time. This prevents
the Fortran global state from being modified while another thread
is trying to run its calculations.

The extension module is still not technically threadsafe, but we are
claiming that our library is when entering from the Python side where
the locks are implemented.