TST/MNT: Add tests for refresh data download util
Add tests to make sure the downloads are triggered appropriately.
This required updating the location to download to a separate temporary
directory for each test. Before we were overwriting the actual test file
and that caused issues in subsequent tests when going to load from that.

Change the condition to only refresh if the data being requested is
before the current time AND after the last date in the file.